refactor(bme sensor): Refactor bme sensor out of main into draft
Draft the environmentaldata trait for sensors with functions that setup
the sensors and read their data to the appropriate protobuf. Possibly
want to change it to modify a generic protobuf rather than create a
protobuf?

Add comments for todos for embassy tasks that scan I2C bus for sensors
and assemble telemetry packets to send.

Add struct to proxy remote errors that do not impl defmt

todo: look into other bme libs in case the IAQ measurements are already
implemented, otherwise just look at the adafruit C++ code that
meshtastic firmware uses to implement this feature for this sensor